1-string_nconcat.c file added

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -1,6 +1,5 @@
 #include "main.h"
 #include <stdlib.h>
-#include <stdio.h>
 
 /**
  * _calloc - allocates memory for an array
@@ -11,7 +10,7 @@
 
 void *_calloc(unsigned int nmemb, unsigned int size)
 {
-	unsigned int i;
+	size_t i;
 	char *ptr_array;
 
 	if (nmemb == 0 || size == 0)
